switch prices to double

+++ src/main/java/org/abondar/industrial/occupancy/data/GuestData.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 @Data
 public class GuestData {
-    private List<Integer> guestPrices;
+    private List<Double> guestPrices;
 }
